Refactor: structure of webserver crate for better readabillity
