+++ rust/scanner-api/src/lib.rs
@@ -1,5 +1,10 @@
-pub mod scan_manager;
-pub mod vt_manager;
+mod scan_manager;
+mod vt_manager;
+
+mod catcher;
+mod fairing;
+mod routes;
 
 pub mod error;
+pub mod manager;
 pub mod webserver;
