+++ rust/scanner-api/src/main.rs
@@ -1,10 +1,8 @@
 use rocket::launch;
-use scanner_api::{
-    scan_manager::DefaultScanManager, vt_manager::DefaultVTManager, webserver::Webserver,
-};
+use scanner_api::webserver::Webserver;
 
 #[launch]
 fn rocket() -> _ {
-    let ws = Webserver::new(DefaultScanManager::new(), DefaultVTManager::new());
+    let ws = Webserver::default();
     ws.run()
 }
